requirements.txt
Babel==2.15.0
python-dotenv==1.0.1
reflex==0.5.9
creditcard==1.0.2